test: pin the streaming-table seam, not the SQL
The dry-run file sat in warehouse/ next to the three tests that mirror
that package's modules, but it had no module behind it and was the only
one to build an Engine. Move it up a level and cut it to the hop nothing
else covers: that the kind the reader observed is the kind the compiler
and the safety rules are handed. The statement matrix stays in
test_compile.py and the refusal rule in test_validation.py.

Also drop a brittle message match from the unknown-scope test and build
the scope table from named groups instead of four spelled-out sets.
